+++ tests/_data/themes/dummy/functions.php
@@ -1,13 +1,11 @@
 <?php
 
 add_action('after_switch_theme', function ($stylesheet) {
-    if ($stylesheet === 'dummy') {
-        update_option('dummy_theme_activated', true);
-    }
+    update_option('dummy_after_switch_theme_called', $stylesheet);
 });
 
-add_action('switch_theme', function () {
-    delete_option('dummy_theme_activated');
+add_action('switch_theme', function ($stylesheet) {
+    update_option('dummy_switch_theme_called', $stylesheet);
 });
 
 function dummy_theme_function()